feat(json.Marshaler): marshal schemas back to json properly.
this makes jsonschema a 2-way street, marshaling decoded schemas
back to json. Not all test edge cases have been worked out, which
we should cover next, but this'll cover lots & lots of use cases.

I'm enforcing the additional constraint that all schema encoding
be lexographic (by using maps for object encoding), because we're
going to be hashing the results of lots of these schemas,
consistent encoding is vital.

word.