feat: return multiple errors on validation call.
This commit introduces a new type: ValError and adjusts the
signature of Validator.Validate to return a slice of ValError
instead of a single error. This change comes with the expectation
that all validators should return the full list of errors that
a given validator detects. Error-free passes can return either
nil or an empty []ValError slice.
This adjustment starts with a basic "just supply the message",
but ValError comes with extra fields that we can iterate upon
to populate in the future.
Having the full list of errors for a given instance makes this
package far more useful, at the expense of forcing implementers
to understand a few extra things. Meh. Worth it.

closes #15