refactor(Validators): export validator structs
So, we've been using this package in the wild for a little while now, and
found a few use cases where having the validator functions themselves exported
is necessary to get introspection into a parsed validator. So, let's export!

In the future we'll hopefully ship a convenience method or two for traversing
validator trees.